test artifact

+++ ci/travis/prepare_artifacts_rpi.sh
@@ -18,14 +18,17 @@ artifacts_structure() {
 	typeKERNEL=( "kernel7" "kernel7l" "kernel" )
 	for index in "${!typeBCM[@]}"; do
 		cd adi_"${typeBCM[$index]}"_defconfig
-		mkdir overlays modules
-		mv ./*.dtbo ./overlays
+		mkdir modules
 		tar -xf rpi_modules.tar.gz -C modules
 		rm rpi_modules.tar.gz
 		mv ./zImage ./"${typeKERNEL[$index]}".img
 		cd ../
 		cp -r ./adi_"${typeBCM[$index]}"_defconfig/* ./${timestamp}
 	done
+	if [ -z "$(ls  ${SOURCE_DIRECTORY}/${timestamp}/*.dtb 2>/dev/null)" ] || [ -z "$(ls  ${SOURCE_DIRECTORY}/${timestamp}/*.dtbo 2>/dev/null)" ]; then
+		echo "Missing one or more required files"
+		exit 1
+	fi
 	tar -C ${SOURCE_DIRECTORY}/${timestamp}/modules -czvf ${SOURCE_DIRECTORY}/${timestamp}/rpi_modules.tar.gz .
 	rm -r ${SOURCE_DIRECTORY}/${timestamp}/modules
 }
@@ -44,6 +47,7 @@ artifacts_swdownloads() {
 	artifacts_structure
 	cd ${SOURCE_DIRECTORY}/${timestamp} || exit 1
 	chmod 600 ${KEY_FILE}
+	BUILD_SOURCEBRANCHNAME="rpi-6.6.y"
 	scp -2 -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -o HostKeyAlgorithms=+ssh-dss \
 		-i ${KEY_FILE} -r *.tar.gz ${DEST_SERVER}/${BUILD_SOURCEBRANCHNAME}
 	md5_modules=($(md5sum rpi_modules.tar.gz| cut -d ' ' -f 1))
